Create README.md
# VBA-challenge

I found this project very challenging.  We are making a list of stocks, finding the difference in price of each from the beginning of the year to the end of the.  
Then we calculate this as a percentage change during the year.  We also add up the volumes of each stock.

We colored the difference column green if it's a positive change and red if it's a negative change.

On the far right side of the sheet, you can see the Greatest % Increase for all the stocks in that list, as well as the Greatest % Decrease.
We also grabbed the stock with the Greatest Total Volume.  

These calculations were all done over three years, 2018 to 2020